Fix gcc-4.9 compilation warning when using -Wstrict-overflow
git-svn-id: http://pugixml.googlecode.com/svn/trunk@957 99668b35-9821-0410-8761-19e4c4f06640

+++ src/pugixml.cpp
@@ -938,7 +938,7 @@ PUGI__NS_BEGIN
 
 			while (data < end)
 			{
-				uint16_t lead = opt_swap::value ? endian_swap(*data) : *data;
+				unsigned int lead = opt_swap::value ? endian_swap(*data) : *data;
 
 				// U+0000..U+D7FF
 				if (lead < 0xD800)
